Rename abstract-item to abstract-item.component
See https://angular.io/styleguide.
