+++ src/client/app/shared/abstract-groups-items/abstract-item.component.ts
